Swagger documentation done
