+++ src/main/java/ar/com/flexibility/examen/config/SwaggerConfig.java
@@ -1,12 +1,14 @@
 package ar.com.flexibility.examen.config;
 
+import java.util.Collections;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
@@ -15,27 +17,24 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
 @EnableSwagger2
 public class SwaggerConfig {
 
-	/**
-     * Publish a bean to generate swagger2 endpoints
-     * @return a swagger configuration bean
-     */
-    @Bean
-    public Docket usersApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .apiInfo(usersApiInfo())
-                .select()
-                .paths(PathSelectors.any())
-                .apis(RequestHandlerSelectors.any())
-                .build()
-                .useDefaultResponseMessages(false);
-    }
-
-   private ApiInfo usersApiInfo() {
-        return new ApiInfoBuilder()
-                .title("Reto Rest")
-                .version("1.0")
-                .license("Apache License Version 2.0")
-                .build();
-    }
+	@Bean
+	public Docket api() {                
+	    return new Docket(DocumentationType.SWAGGER_2)          
+	      .select()
+	      .apis(RequestHandlerSelectors.basePackage("ar.com.flexibility.examen.app"))
+	      .paths(PathSelectors.any())
+	      .build()
+	      .apiInfo(apiInfo());
+	}
+	 
+	private ApiInfo apiInfo() {
+	    return new ApiInfo(
+	      "Reto REST API", 
+	      "Reto servicios REST API - Flexibility.", 
+	      "V1.0", 
+	      "Flexibility", 
+	      new Contact("Juan Pablo", "www.reto-rest.com", "juanpablo@company.com"), 
+	      "License of API", "API license URL", Collections.emptyList());
+	}
    
 }
